requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.3.2
apipkg==1.5               # via execnet
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, pytest
black==18.9b0
certifi==2019.11.28       # via sentry-sdk
click==7.0                # via black
coverage==5.0.3
entrypoints==0.3          # via flake8
execnet==1.7.1            # via pytest-xdist
flake8-import-order==0.18.1  # via zimports
flake8==3.7.9
grpcio-tools==1.26.0
grpcio==1.26.0
mako==1.1.1               # via alembic
markupsafe==1.1.1         # via mako
mccabe==0.6.1             # via flake8
more-itertools==8.1.0     # via pytest
mysqlclient==1.4.6
packaging==20.1           # via pytest
passlib==1.7.2
phonenumbers==8.11.1
pluggy==0.13.1            # via pytest
prometheus-client==0.7.1  # via py-grpc-prometheus
protobuf==3.11.2          # via grpcio-tools
py-grpc-prometheus==0.2.0
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8, flake8-import-order
pyflakes==2.1.1           # via flake8, zimports
pyjwt==1.7.1
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1
pytest-forked==1.1.3      # via pytest-xdist
pytest-xdist==1.31.0
pytest==5.3.4
python-dateutil==2.8.1    # via alembic
python-editor==1.0.4      # via alembic
sentry-sdk==0.14.1
six==1.14.0               # via grpcio, packaging, protobuf, pytest-xdist, python-dateutil, sqlalchemy-utils
sqlalchemy-utils==0.36.1
sqlalchemy==1.3.12
toml==0.10.0              # via black
urllib3==1.25.8           # via sentry-sdk
wcwidth==0.1.8            # via pytest
zimports==0.2.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
